// dpr/types/api.d.ts
export interface components {
  schemas: {
    ErrorResponse: {
      /** Format: int32 */
      status: number
      /** Format: int32 */
      errorCode?: number
      userMessage?: string
      developerMessage?: string
      moreInfo?: string
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}